Moved last files in api to dedicated folders
