+++ nextjs-blog/pages/api/project/getAllProject.js
@@ -1,5 +1,5 @@
 
-import clientPromise from "../../lib/mongodb";
+import clientPromise from "../../../lib/mongodb";
 
 export default async (req, res) => {
    try {
